Move the Eigen source out of the module
While having a local copy of the Eigen header files stored in this module seems
is stable, it introduces problems for other C++ programs that need to
use Eigen. They must also include at very least the datatypes. So
instead, we should install Eigen in the $MORLOC_HOME/include directory
and include the file from there.